Refactor player and immersive home into memoized components
