+++ app/src/androidTest/java/com/rpeters/jellyfin/ui/screens/ImmersiveHomeScreenTest.kt
@@ -1,12 +1,14 @@
 package com.rpeters.jellyfin.ui.screens
 
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.ui.test.assertDoesNotExist
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.rpeters.jellyfin.ui.screens.home.HomeContentLists
+import com.rpeters.jellyfin.ui.screens.home.MobileExpressiveHomeContent
 import com.rpeters.jellyfin.ui.theme.JellyfinAndroidTheme
 import com.rpeters.jellyfin.ui.viewmodel.MainAppState
 import io.mockk.mockk
