+++ app/src/main/java/com/rpeters/jellyfin/ui/screens/ImmersiveHomeScreen.kt
@@ -484,465 +484,3 @@ private fun ImmersiveHomeContent(
         }
     }
 }
-
-@VisibleForTesting
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-internal fun MobileExpressiveHomeContent(
-    appState: MainAppState,
-    contentLists: HomeContentLists,
-    getImageUrl: (BaseItemDto) -> String?,
-    getBackdropUrl: (BaseItemDto) -> String?,
-    getSeriesImageUrl: (BaseItemDto) -> String?,
-    onItemClick: (BaseItemDto) -> Unit,
-    onItemLongPress: (BaseItemDto) -> Unit,
-    onLibraryClick: (BaseItemDto) -> Unit,
-    viewingMood: String?,
-    listState: LazyListState,
-    contentPadding: PaddingValues,
-    bottomSpacing: androidx.compose.ui.unit.Dp,
-    animatedVisibilityScope: androidx.compose.animation.AnimatedVisibilityScope? = null,
-    modifier: Modifier = Modifier,
-) {
-    val unknownText = stringResource(id = R.string.unknown)
-    val libraryRows = remember(appState.libraries, appState.itemsByLibrary) {
-        appState.libraries.mapNotNull { library ->
-            if (library.toLibraryTypeOrNull() == LibraryType.STUFF) return@mapNotNull null
-            val libraryId = library.id.toString()
-            val items = appState.itemsByLibrary[libraryId]
-                .orEmpty()
-                .sortedByDescending { it.dateCreated }
-                .take(10)
-            if (items.isEmpty()) null else library to items
-        }
-    }
-
-    LazyColumn(
-        state = listState,
-        modifier = modifier,
-        contentPadding = PaddingValues(
-            start = 0.dp,
-            top = contentPadding.calculateTopPadding(),
-            end = 0.dp,
-            bottom = contentPadding.calculateBottomPadding() + bottomSpacing,
-        ),
-        verticalArrangement = Arrangement.spacedBy(20.dp),
-    ) {
-        // Replace wonky carousel with high-impact Recently Added Movies (Last 5)
-        val heroMovies = contentLists.recentMovies.take(5)
-        if (heroMovies.isNotEmpty() || appState.isLoading) {
-            item(key = "recently_added_hero", contentType = "hero_row") {
-                if (heroMovies.isNotEmpty()) {
-                    val featuredItems = remember(heroMovies, unknownText) {
-                        heroMovies.map { movie ->
-                            CarouselItem(
-                                id = movie.id.toString(),
-                                title = movie.name ?: unknownText,
-                                subtitle = movie.productionYear?.toString() ?: "",
-                                imageUrl = getBackdropUrl(movie) ?: getImageUrl(movie) ?: "",
-                                type = MediaType.MOVIE,
-                            )
-                        }
-                    }
-
-                    ImmersiveHeroCarousel(
-                        items = featuredItems,
-                        onItemClick = { selected ->
-                            heroMovies.firstOrNull { it.id.toString() == selected.id }?.let(onItemClick)
-                        },
-                        onPlayClick = { selected ->
-                            heroMovies.firstOrNull { it.id.toString() == selected.id }?.let(onItemClick)
-                        },
-                        modifier = Modifier.fillMaxWidth(),
-                    )
-                } else {
-                    // Placeholder while loading to prevent layout jump
-                    Box(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(com.rpeters.jellyfin.ui.theme.ImmersiveDimens.HeroHeightPhone),
-                    )
-                }
-            }
-        }
-
-        item(key = "libraries", contentType = "libraries") {
-            LibraryNavigationCarousel(
-                libraries = appState.libraries,
-                getImageUrl = getImageUrl,
-                onLibraryClick = onLibraryClick,
-            )
-        }
-
-        if (!viewingMood.isNullOrBlank()) {
-            item(key = "viewing_mood", contentType = "viewing_mood") {
-                ViewingMoodWidget(
-                    viewingMood = viewingMood,
-                    modifier = Modifier.padding(horizontal = 16.dp),
-                )
-            }
-        }
-
-        if (contentLists.continueWatching.isNotEmpty()) {
-            item(key = "continue_watching", contentType = "continue_watching") {
-                ContinueWatchingSection(
-                    items = contentLists.continueWatching,
-                    getImageUrl = { item -> getBackdropUrl(item) ?: getSeriesImageUrl(item) ?: getImageUrl(item) },
-                    onItemClick = onItemClick,
-                    onItemLongPress = onItemLongPress,
-                    cardWidth = 320.dp,
-                )
-            }
-        }
-
-        if (contentLists.recentEpisodes.isNotEmpty()) {
-            item(key = "next_up", contentType = "next_up") {
-                PosterRowSection(
-                    title = stringResource(id = R.string.home_recently_added_tv_episodes),
-                    items = contentLists.recentEpisodes,
-                    getImageUrl = { item -> getSeriesImageUrl(item) ?: getImageUrl(item) },
-                    onItemClick = onItemClick,
-                    onItemLongPress = onItemLongPress,
-                    cardWidth = 200.dp,
-                )
-            }
-        }
-
-        if (contentLists.recentVideos.isNotEmpty()) {
-            item(key = "recent_stuff", contentType = "recent_stuff") {
-                MediaRowSection(
-                    title = stringResource(id = R.string.home_recently_added_stuff),
-                    items = contentLists.recentVideos,
-                    getImageUrl = { item -> getBackdropUrl(item) ?: getImageUrl(item) },
-                    onItemClick = onItemClick,
-                    onItemLongPress = onItemLongPress,
-                )
-            }
-        }
-
-        items(
-            items = libraryRows,
-            key = { (library, _) -> library.id.toString() },
-            contentType = { "library_recent_row" },
-        ) { (library, items) ->
-            LibraryRecentSection(
-                library = library,
-                items = items,
-                getImageUrl = { item ->
-                    when (item.type) {
-                        BaseItemKind.EPISODE -> getSeriesImageUrl(item) ?: getImageUrl(item)
-                        BaseItemKind.SERIES -> getSeriesImageUrl(item) ?: getBackdropUrl(item) ?: getImageUrl(item)
-                        BaseItemKind.VIDEO -> getBackdropUrl(item) ?: getImageUrl(item)
-                        else -> getImageUrl(item)
-                    }
-                },
-                onLibraryClick = onLibraryClick,
-                onItemClick = onItemClick,
-                onItemLongPress = onItemLongPress,
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun LibraryNavigationCarousel(
-    libraries: List<BaseItemDto>,
-    getImageUrl: (BaseItemDto) -> String?,
-    onLibraryClick: (BaseItemDto) -> Unit,
-) {
-    if (libraries.isEmpty()) return
-
-    Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-        HomeSectionTitle(
-            title = stringResource(R.string.home_libraries),
-            modifier = Modifier.padding(top = 8.dp),
-        )
-
-        val carouselState = rememberCarouselState { libraries.size }
-        androidx.compose.material3.carousel.HorizontalUncontainedCarousel(
-            state = carouselState,
-            itemWidth = 220.dp, // Increased from 176dp
-            itemSpacing = 16.dp, // Increased spacing
-            contentPadding = PaddingValues(horizontal = 16.dp),
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(220.dp), // Match item width for balanced look
-        ) { index ->
-            val library = libraries[index]
-            LibraryExpressiveCard(
-                library = library,
-                imageUrl = getImageUrl(library),
-                onClick = { onLibraryClick(library) },
-                modifier = Modifier.fillMaxSize(),
-            )
-        }
-    }
-}
-
-@Composable
-private fun LibraryRecentSection(
-    library: BaseItemDto,
-    items: List<BaseItemDto>,
-    getImageUrl: (BaseItemDto) -> String?,
-    onLibraryClick: (BaseItemDto) -> Unit,
-    onItemClick: (BaseItemDto) -> Unit,
-    onItemLongPress: (BaseItemDto) -> Unit,
-) {
-    val haptics = com.rpeters.jellyfin.ui.utils.rememberExpressiveHaptics()
-
-    Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 16.dp),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Column(
-                modifier = Modifier.weight(1f),
-                verticalArrangement = Arrangement.spacedBy(2.dp),
-            ) {
-                Text(
-                    text = library.name ?: "",
-                    style = MaterialTheme.typography.titleLarge,
-                    fontWeight = FontWeight.Bold,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-                Text(
-                    text = stringResource(R.string.library_recently_added_section),
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-            }
-            FilledTonalButton(
-                onClick = {
-                    haptics.lightClick()
-                    onLibraryClick(library)
-                },
-            ) {
-                Text(text = stringResource(R.string.open))
-            }
-        }
-
-        when (library.toLibraryTypeOrNull()) {
-            LibraryType.STUFF, LibraryType.MUSIC -> HomeLibraryMediaRow(
-                items = items,
-                getImageUrl = getImageUrl,
-                onItemClick = onItemClick,
-                onItemLongPress = onItemLongPress,
-                cardWidth = 320.dp, // Increased from 240dp
-            )
-            else -> HomeLibraryPosterRow(
-                items = items,
-                getImageUrl = getImageUrl,
-                onItemClick = onItemClick,
-                onItemLongPress = onItemLongPress,
-                cardWidth = 200.dp, // Increased from 164dp
-            )
-        }
-    }
-}
-
-@Composable
-private fun HomeLibraryPosterRow(
-    items: List<BaseItemDto>,
-    getImageUrl: (BaseItemDto) -> String?,
-    onItemClick: (BaseItemDto) -> Unit,
-    onItemLongPress: (BaseItemDto) -> Unit,
-    cardWidth: androidx.compose.ui.unit.Dp,
-) {
-    LazyRow(
-        contentPadding = PaddingValues(horizontal = 16.dp),
-        horizontalArrangement = Arrangement.spacedBy(16.dp),
-    ) {
-        items(items = items, key = { it.getItemKey() }) { item ->
-            PosterMediaCard(
-                item = item,
-                getImageUrl = getImageUrl,
-                onClick = onItemClick,
-                onLongPress = onItemLongPress,
-                cardWidth = cardWidth,
-                showTitle = true,
-                showMetadata = true,
-                titleMinLines = 2,
-            )
-        }
-    }
-}
-
-@Composable
-private fun HomeLibraryMediaRow(
-    items: List<BaseItemDto>,
-    getImageUrl: (BaseItemDto) -> String?,
-    onItemClick: (BaseItemDto) -> Unit,
-    onItemLongPress: (BaseItemDto) -> Unit,
-    cardWidth: androidx.compose.ui.unit.Dp,
-) {
-    LazyRow(
-        contentPadding = PaddingValues(horizontal = 16.dp),
-        horizontalArrangement = Arrangement.spacedBy(16.dp),
-    ) {
-        items(items = items, key = { it.getItemKey() }) { item ->
-            MediaCard(
-                item = item,
-                getImageUrl = { getImageUrl(item) },
-                onClick = onItemClick,
-                onLongPress = onItemLongPress,
-                cardWidth = cardWidth,
-            )
-        }
-    }
-}
-
-@Composable
-private fun LibraryExpressiveCard(
-    library: BaseItemDto,
-    imageUrl: String?,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val haptics = com.rpeters.jellyfin.ui.utils.rememberExpressiveHaptics()
-    val sharedTransitionScope = com.rpeters.jellyfin.ui.navigation.LocalSharedTransitionScope.current
-    val animatedVisibilityScope = com.rpeters.jellyfin.ui.navigation.LocalAnimatedVisibilityScope.current
-    val libraryId = library.id.toString()
-
-    val sharedElementModifier = if (sharedTransitionScope != null && animatedVisibilityScope != null) {
-        with(sharedTransitionScope) {
-            Modifier.sharedElement(
-                rememberSharedContentState(key = "library_$libraryId"),
-                animatedVisibilityScope = animatedVisibilityScope,
-            )
-        }
-    } else {
-        Modifier
-    }
-
-    ElevatedCard(
-        onClick = {
-            haptics.lightClick()
-            onClick()
-        },
-        modifier = modifier
-            .then(sharedElementModifier),
-        shape = RoundedCornerShape(28.dp),
-        colors = CardDefaults.elevatedCardColors(
-            containerColor = MaterialTheme.colorScheme.surfaceContainerHigh.copy(alpha = 0.82f),
-        ),
-    ) {
-        Box(modifier = Modifier.fillMaxSize()) {
-            HeroImageWithGradient(
-                imageUrl = imageUrl,
-                contentDescription = library.name ?: "Library",
-                modifier = Modifier.fillMaxSize(),
-            )
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(16.dp),
-            ) {
-                Surface(
-                    shape = CircleShape,
-                    color = MaterialTheme.colorScheme.surface.copy(alpha = 0.82f),
-                ) {
-                    Icon(
-                        imageVector = library.toLibraryTypeOrNull()?.icon ?: Icons.Default.Folder,
-                        contentDescription = null,
-                        modifier = Modifier.padding(10.dp).size(20.dp),
-                    )
-                }
-
-                Column(
-                    modifier = Modifier.align(Alignment.BottomStart),
-                    verticalArrangement = Arrangement.spacedBy(4.dp),
-                ) {
-                    Surface(
-                        shape = RoundedCornerShape(20.dp),
-                        color = MaterialTheme.colorScheme.surfaceContainerHighest.copy(alpha = 0.86f),
-                        tonalElevation = 2.dp,
-                    ) {
-                        Column(
-                            modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp),
-                            verticalArrangement = Arrangement.spacedBy(2.dp),
-                        ) {
-                            Text(
-                                text = library.name ?: "Library",
-                                style = MaterialTheme.typography.titleMedium,
-                                color = MaterialTheme.colorScheme.onSurface,
-                                fontWeight = FontWeight.Bold,
-                                maxLines = 2,
-                                overflow = TextOverflow.Ellipsis,
-                            )
-                            Text(
-                                text = library.collectionType?.toString()?.replace("_", " ") ?: "Collection",
-                                style = MaterialTheme.typography.bodySmall,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                maxLines = 1,
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-private fun BaseItemDto.toLibraryTypeOrNull(): LibraryType? = when (collectionType) {
-    CollectionType.MOVIES -> LibraryType.MOVIES
-    CollectionType.TVSHOWS -> LibraryType.TV_SHOWS
-    CollectionType.MUSIC -> LibraryType.MUSIC
-    CollectionType.HOMEVIDEOS, CollectionType.BOOKS -> LibraryType.STUFF
-    else -> when (collectionType?.toString()?.lowercase()?.replace(" ", "")) {
-        "movies" -> LibraryType.MOVIES
-        "tvshows" -> LibraryType.TV_SHOWS
-        "music" -> LibraryType.MUSIC
-        else -> LibraryType.STUFF
-    }
-}
-
-/**
- * Viewing mood widget with AI-powered mood analysis
- */
-@Composable
-private fun ViewingMoodWidget(
-    viewingMood: String,
-    onMoodClick: () -> Unit = {},
-    modifier: Modifier = Modifier,
-) {
-    ExpressiveContentCard(
-        onClick = onMoodClick,
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(vertical = Dimens.Spacing8),
-        containerColor = MaterialTheme.colorScheme.tertiaryContainer,
-        contentColor = MaterialTheme.colorScheme.onTertiaryContainer,
-    ) {
-        Row(
-            modifier = Modifier.padding(ImmersiveDimens.SpacingContentPadding),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(Dimens.Spacing12),
-        ) {
-            Icon(
-                imageVector = Icons.Default.AutoAwesome,
-                contentDescription = null,
-                tint = MaterialTheme.colorScheme.onTertiaryContainer,
-                modifier = Modifier.size(32.dp),
-            )
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = "Your Viewing Mood",
-                    style = MaterialTheme.typography.labelLarge,
-                    fontWeight = FontWeight.Bold,
-                    color = MaterialTheme.colorScheme.onTertiaryContainer,
-                )
-                Spacer(modifier = Modifier.height(4.dp))
-                Text(
-                    text = viewingMood,
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onTertiaryContainer,
-                )
-            }
-        }
-    }
-}
